daemon struct, sockets + havens, cleanup
