+++ src/daemon/global_rpc/transport.rs
@@ -6,9 +6,9 @@ use futures_util::{future, FutureExt};
 use nanorpc::{JrpcRequest, JrpcResponse, RpcTransport};
 use smol::Timer;
 
-use crate::daemon::{
-    n2r_socket::{Endpoint, N2rSocket},
-    DaemonContext,
+use crate::{
+    daemon::context::DaemonContext,
+    sockets::{n2r_socket::N2rSocket, socket::Endpoint},
 };
 
 use super::GLOBAL_RPC_DOCK;
