+++ src/sockets/n2r_socket.rs
@@ -1,14 +1,14 @@
-use std::{fmt::Display, str::FromStr, sync::Arc};
+use std::sync::Arc;
 
-use crate::daemon::DaemonContext;
 use bytes::Bytes;
 use earendil_crypt::{Fingerprint, IdentitySecret};
 use earendil_packet::{Dock, Message};
 use rand::Rng;
-use serde::{Deserialize, Serialize};
 use smol::channel::Receiver;
 
-use super::socket::{SocketRecvError, SocketSendError};
+use crate::daemon::context::DaemonContext;
+
+use super::socket::{Endpoint, SocketRecvError, SocketSendError};
 
 pub struct N2rSocket {
     ctx: DaemonContext,
@@ -103,39 +103,5 @@ impl Drop for BoundDock {
         self.ctx
             .socket_recv_queues
             .remove(&Endpoint::new(self.fp, self.dock));
-    }
-}
-
-#[derive(Copy, Clone, Deserialize, Serialize, Hash, Debug, PartialEq, PartialOrd, Ord, Eq)]
-pub struct Endpoint {
-    pub fingerprint: Fingerprint,
-    pub dock: Dock,
-}
-
-impl Endpoint {
-    pub fn new(fingerprint: Fingerprint, dock: Dock) -> Endpoint {
-        Endpoint { fingerprint, dock }
-    }
-}
-
-impl Display for Endpoint {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        write!(f, "{}:{}", self.fingerprint, self.dock)
-    }
-}
-
-impl FromStr for Endpoint {
-    type Err = anyhow::Error;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let elems: Vec<&str> = s.split(":").collect();
-        if elems.len() != 2 {
-            return Err(anyhow::anyhow!(
-                "Wrong endpoint format! Endpoint format should be fingerprint:dock"
-            ));
-        }
-        let fp = Fingerprint::from_str(elems[0])?;
-        let dock = u32::from_str(elems[1])?;
-        Ok(Endpoint::new(fp, dock))
     }
 }
