+++ src/sockets/socket.rs
@@ -1,16 +1,14 @@
+use std::{fmt::Display, str::FromStr};
+
 use bytes::Bytes;
 use earendil_crypt::{Fingerprint, IdentitySecret};
 use earendil_packet::Dock;
 use serde::{Deserialize, Serialize};
 use thiserror::Error;
 
-use crate::control_protocol::SendMessageError;
+use crate::{control_protocol::SendMessageError, daemon::context::DaemonContext};
 
-use super::{
-    haven_socket::HavenSocket,
-    n2r_socket::{Endpoint, N2rSocket},
-    DaemonContext,
-};
+use super::{haven_socket::HavenSocket, n2r_socket::N2rSocket};
 
 pub struct Socket {
     inner: InnerSocket,
@@ -83,4 +81,38 @@ pub enum SocketRecvError {
     N2rRecvError,
     #[error("improperly formatted inner haven message")]
     HavenMsgBadFormat,
+}
+
+#[derive(Copy, Clone, Deserialize, Serialize, Hash, Debug, PartialEq, PartialOrd, Ord, Eq)]
+pub struct Endpoint {
+    pub fingerprint: Fingerprint,
+    pub dock: Dock,
+}
+
+impl Endpoint {
+    pub fn new(fingerprint: Fingerprint, dock: Dock) -> Endpoint {
+        Endpoint { fingerprint, dock }
+    }
+}
+
+impl Display for Endpoint {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        write!(f, "{}:{}", self.fingerprint, self.dock)
+    }
+}
+
+impl FromStr for Endpoint {
+    type Err = anyhow::Error;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        let elems: Vec<&str> = s.split(":").collect();
+        if elems.len() != 2 {
+            return Err(anyhow::anyhow!(
+                "Wrong endpoint format! Endpoint format should be fingerprint:dock"
+            ));
+        }
+        let fp = Fingerprint::from_str(elems[0])?;
+        let dock = u32::from_str(elems[1])?;
+        Ok(Endpoint::new(fp, dock))
+    }
 }
